examples/lightstep-llamaindex/requirements.txt
Flask
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-launcher
llama-index
requests